添加 播放历史记录

+++ app/src/main/java/tech/shmy/dd_app/entity/VideoEntity.java
@@ -4,7 +4,7 @@ import java.util.Date;
 import java.util.List;
 
 public class VideoEntity {
-    public int id;
+    public long id;
     public String pic;
     public String name;
     public String actor;
@@ -15,8 +15,8 @@ public class VideoEntity {
     public String year;
     public List<LabelEntity> lable;
     public List<ResourceEntity> resources;
-    public int pid;
-    public int views;
+    public long pid;
+    public long views;
     public Date created_at;
     public Date updated_at;
     public Date deleted_at;
